+++ python/src/pypalmsens/_data/shared.py
