+++ python/src/pypalmsens/_instruments/instrument_manager_async.py
@@ -21,9 +21,9 @@ from .._methods import (
     cr_string_to_enum,
 )
 from ..data import Measurement
-from ._common import Instrument, create_future, firmware_warning
 from .callback import Callback, CallbackStatus, Status
 from .measurement_manager_async import MeasurementManagerAsync
+from .shared import Instrument, create_future, firmware_warning
 
 WINDOWS = sys.platform == 'win32'
 LINUX = not WINDOWS
