+++ python/src/pypalmsens/_instruments/instrument_pool_async.py
@@ -4,8 +4,8 @@ import asyncio
 from typing import TYPE_CHECKING, Any, Awaitable, Callable, Sequence
 
 from .._methods import BaseTechnique
-from ._common import Instrument
 from .instrument_manager_async import InstrumentManagerAsync
+from .shared import Instrument
 
 if TYPE_CHECKING:
     from .._data.measurement import Measurement
