+++ python/src/pypalmsens/_instruments/shared.py
