+++ python/src/pypalmsens/_methods/settings.py
@@ -8,8 +8,9 @@ from PalmSens import MuxMethod as PSMuxMethod
 from pydantic import Field
 from typing_extensions import override
 
-from .._shared import single_to_double
-from ._shared import (
+from .._helpers import single_to_double
+from .base import BaseSettings
+from .shared import (
     AllowedCurrentRanges,
     AllowedPotentialRanges,
     convert_bools_to_int,
@@ -19,7 +20,6 @@ from ._shared import (
     pr_enum_to_string,
     pr_string_to_enum,
 )
-from .base import BaseSettings
 
 
 class CurrentRange(BaseSettings):
