+++ python/src/pypalmsens/data.py
@@ -2,13 +2,13 @@
 
 from __future__ import annotations
 
-from ._data._shared import ArrayType
 from ._data.curve import Curve
 from ._data.data_array import DataArray
 from ._data.dataset import DataSet
 from ._data.eisdata import EISData
 from ._data.measurement import DeviceInfo, Measurement
 from ._data.peak import Peak
+from ._data.shared import ArrayType
 from ._instruments.callback import (
     CallbackData,
     CallbackDataEIS,
