+++ python/tests/test_dataset.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from pypalmsens._data._shared import ArrayType
+from pypalmsens._data.shared import ArrayType
 from pypalmsens.data import DataArray
